requirements.txt
alembic==1.8.1
bcrypt==3.2.2
blinker==1.5
Bootstrap-Flask==2.0.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.0
WTForms==3.0.1
uwsgi